- Update enum name to match convention - Move tests to a more appropriate location under integration-tests/commands
