+++ test/integration-tests/testexplorer/TestExplorerIntegration.test.ts
@@ -42,7 +42,7 @@ import {
     reduceTestItemChildren,
 } from "../../../src/TestExplorer/TestUtils";
 import { runnableTag } from "../../../src/TestExplorer/TestDiscovery";
-import { COMMANDS } from "../../../src/commands";
+import { Commands } from "../../../src/commands";
 
 suite("Test Explorer Suite", function () {
     const MAX_TEST_RUN_TIME_MINUTES = 5;
@@ -305,7 +305,7 @@ suite("Test Explorer Suite", function () {
                     // Stub the showInputBox method to return the input text
                     windowMock.showInputBox.resolves(`${numIterations}`);
 
-                    vscode.commands.executeCommand(COMMANDS.runTestsMultipleTimes, testItems[0]);
+                    vscode.commands.executeCommand(Commands.runTestsMultipleTimes, testItems[0]);
 
                     const testRun = await eventPromise(testExplorer.onCreateTestRun);
 
@@ -420,7 +420,7 @@ suite("Test Explorer Suite", function () {
                     // Stub the showInputBox method to return the input text
                     windowMock.showInputBox.resolves(`${numIterations}`);
 
-                    vscode.commands.executeCommand(COMMANDS.runTestsMultipleTimes, testItems[0]);
+                    vscode.commands.executeCommand(Commands.runTestsMultipleTimes, testItems[0]);
 
                     const testRun = await eventPromise(testExplorer.onCreateTestRun);
